Little refactoring and small fixes (Refs #468)

+++ cuckoo.py
@@ -85,18 +85,17 @@ def cuckoo_clean():
 
     # Check if MongoDB reporting is enabled and drop that if it is.
     cfg = Config("reporting")
-    if cfg.mongodb:
-        if cfg.mongodb.enabled:
-            from pymongo import MongoClient
-            host = cfg.mongodb.host
-            port = cfg.mongodb.port
-            mdb = cfg.mongodb.db
-            try:
-                conn = MongoClient(host, port)
-                conn.drop_database(mdb)
-                conn.disconnect()
-            except:
-                log.warning("Unable to drop MongoDB Database: %s", mdb)
+    if cfg.mongodb and cfg.mongodb.enabled:
+        from pymongo import MongoClient
+        host = cfg.mongodb.get("host", "127.0.0.1")
+        port = cfg.mongodb.get("port", 27017)
+        mdb = cfg.mongodb.get("db", "cuckoo")
+        try:
+            conn = MongoClient(host, port)
+            conn.drop_database(mdb)
+            conn.disconnect()
+        except:
+            log.warning("Unable to drop MongoDB database: %s", mdb)
 
     # Paths to clean
     paths = [
